refactor(solve-page): ctrl+enter to submit ExtraPrompt
